requirements.txt
python-dotenv
aiogram
motor
pydantic
pymongo